added c solution for: 151, 152, 153, 155

+++ algorithms/2_medium/F/00153_find_minimum_in_rotated_sorted_array/two_pointers.rs
@@ -1,18 +1,17 @@
 impl Solution {
     pub fn find_min(nums: Vec<i32>) -> i32 {
-
-        let mut left = 0_usize;
-        let mut right = nums.len() - 1;
+        let mut left = 0_isize;
+        let mut right = nums.len() as isize - 1;
         let mut min_val = i32::MAX;
 
         while left <= right {
             let mid = (right + left) / 2;
 
-            if nums[left] <= nums[mid] {
-                min_val = min_val.min(nums[left]);
+            if nums[left as usize] <= nums[mid as usize] {
+                min_val = min_val.min(nums[left as usize]);
                 left = mid + 1;
             } else {
-                min_val = min_val.min(nums[mid]);
+                min_val = min_val.min(nums[mid as usize]);
                 right = mid - 1;
             }
         }
